+++ Bot/Libs/economy_utils/eco_main.py
@@ -1,16 +1,21 @@
 import asyncio
-from typing import Optional
+import uuid
+from typing import Optional, Union
 
 import motor.motor_asyncio
 import uvloop
 from beanie import init_beanie
+from discord.utils import utcnow
+from sqlalchemy.engine.row import Row
 
 from .marketplace_models import ItemAuthProject, MarketplaceModel, PurchaseProject
+from .user_inv import KumikoUserInvUtils
 
 
 class KumikoEcoUtils:
     def __init__(self):
         self.self = self
+        self.userInvUtils = KumikoUserInvUtils()
 
     # Used: 1
     async def ins(
@@ -212,5 +217,106 @@ class KumikoEcoUtils:
         await MarketplaceModel.find(
             MarketplaceModel.uuid == uuid, MarketplaceModel.owner == owner
         ).delete()
+
+    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
+
+    # -------------------------------- #
+    # ! NEW PURCHASE SYSTEM COROUTINES #
+    # -------------------------------- #
+
+    async def getRequestedPurchaseItem(self, name: str, uri: str):
+        """Gets the first or none item that matches the nam
+
+        This is used to first search up the item to purchase from the Marketplace
+
+        Args:
+            name (str): The name of the item to search up
+            uri (str): MongoDB Connection URI
+        """
+        clientPurchase = motor.motor_asyncio.AsyncIOMotorClient(uri)
+        await init_beanie(
+            database=clientPurchase.kumiko_marketplace,
+            document_models=[MarketplaceModel],
+        )
+        return await MarketplaceModel.find(
+            MarketplaceModel.name == name
+        ).first_or_none()
+
+    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
+
+    async def setCurrentStock(self, uuid: str, amount: int, uri: str) -> None:
+        """Sets the current stock amount for the requested item (checked using the UUID provided)
+
+        Args:
+            uuid (str): Requested Item UUID
+            amount (int): The amount to set to. This is also commonly used to set it to be 0
+            uri (str): MongoDB Connection URI
+        """
+        clientUpdate = motor.motor_asyncio.AsyncIOMotorClient(uri)
+        await init_beanie(
+            database=clientUpdate.kumiko_marketplace,
+            document_models=[MarketplaceModel],
+        )
+        marketplaceItem = await MarketplaceModel.find_one(MarketplaceModel.uuid == uuid)
+        marketplaceItem.amount = amount
+        await marketplaceItem.save()
+
+    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
+
+    async def purchaseItem(
+        self,
+        user_inv: Union[Row, None],
+        requested_item: Union[MarketplaceModel, None],
+        current_stock: int,
+        requested_amount: int,
+        user_id: int,
+        mongo_uri: str,
+        postgres_uri: str,
+    ):
+        """Makes the purchase for the requested item
+
+        This should be used with purchasing the item
+
+        Args:
+            user_inv (Union[Row, None]): The object (can be None) of the user's inventory
+            requested_item (Union[MarketplaceModel, None]): Object (can be None) of the requested item
+            current_stock (int): The current stock that the requested item has
+            requested_amount (int): The amount that is requested to be purchased for
+            user_id (int): The Discord user's ID
+            mongo_uri (str): MongoDB Connection URI
+            postgres_uri (str): PostgreSQL Connection URI
+        """
+        requestedItemUUID = dict(requested_item)["uuid"]
+        returnStr = ""
+        if user_inv is None:
+            await self.updateItemAmount(
+                uuid=requestedItemUUID, amount=current_stock, uri=mongo_uri
+            )
+            await self.userInvUtils.insertItem(
+                user_uuid=str(uuid.uuid4()),
+                user_id=user_id,
+                date_acquired=utcnow().isoformat(),
+                name=dict(requested_item)["name"],
+                description=dict(requested_item)["description"],
+                amount=requested_amount,
+                uri=postgres_uri,
+            )
+            returnStr = f"Purchased {requested_amount} {dict(requested_item)['name']} for {dict(requested_item)['price']} coins. Also added {dict(requested_item)['name']} to your inventory."
+        else:
+            amountAddingToUser = int(dict(user_inv[0])["amount"]) + int(
+                requested_amount
+            )
+            await self.updateItemAmount(
+                uuid=requestedItemUUID, amount=current_stock, uri=mongo_uri
+            )
+            await self.userInvUtils.updateItemAmount(
+                user_id=user_id,
+                uuid=requestedItemUUID,
+                amount=amountAddingToUser,
+                uri=postgres_uri,
+            )
+            returnStr = f"Purchased {requested_amount} {dict(requested_item)['name']} for {dict(requested_item)['price']} coins."
+
+        return returnStr
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())